+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Added: adding listeners via reflection through the `map()` method
+
 ## 0.1.0
 
 - Initial version